+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Use the new docker-push image format for Gitlab pipelines
 * Refine the check for when 3rd-party licenses are missing
 * Add Python 'Invoke' for task management
+* Use a pod template for generated injection and cleanup pods